# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Visual Diagram Gen API",
    "version": "1.0.0",
    "description": "Backend API for the AI explainer animation tool. This document describes the public HTTP contract used by the frontend and MCP tooling."
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development API" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "array", "items": {} }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
...
      "Component": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "displayName": { "type": "string" },
          "description": { "type": "string" },
          "propsSchema": { "type": "object" },
          "defaultProps": { "type": "object" },
          "enterStyles": { "type": "array", "items": { "type": "string" } },
          "exitStyles": { "type": "array", "items": { "type": "string" } },
          "colorProps": { "type": "array", "items": { "type": "string" } },
          "refProps": { "type": "array", "items": { "type": "string" } },
          "assetProps": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "description",
          "propsSchema",
          "defaultProps",
          "enterStyles",
          "exitStyles",
          "colorProps",
          "refProps",
          "assetProps",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Issue": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["path", "code", "message"]
      },
...
      "ComponentWriteRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "displayName": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "minLength": 1 },
          "propsSchema": { "type": "object" },
          "defaultProps": { "type": "object" },
          "enterStyles": { "type": "array", "items": { "type": "string" } },
          "exitStyles": { "type": "array", "items": { "type": "string" } },
          "colorProps": { "type": "array", "items": { "type": "string" } },
          "refProps": { "type": "array", "items": { "type": "string" } },
          "assetProps": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "name",
          "displayName",
          "description",
          "propsSchema",
          "defaultProps",
          "enterStyles",
          "exitStyles",
          "colorProps",
          "refProps",
          "assetProps"
        ]
      },
      "ItemsComponentResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Component" }
          }
        },
        "required": ["items"]
      },
...
      "GenerationRequest": {
        "type": "object",
        "properties": {
          "transcriptId": { "type": "string", "format": "uuid" },
          "instructions": { "type": "string", "maxLength": 20000 },
          "save": { "type": "boolean" },
          "provider": {
            "type": "string",
            "enum": ["gemini", "manual"]
          },
          "apiKey": { "type": "string", "writeOnly": true, "maxLength": 500 },
          "model": { "type": "string", "maxLength": 200 }
        },
        "required": ["transcriptId"]
      },
      "ManualPromptRequest": {
        "type": "object",
        "properties": {
          "transcriptId": { "type": "string", "format": "uuid" },
          "instructions": { "type": "string", "maxLength": 20000 }
        },
        "required": ["transcriptId"]
      },
      "ManualSubmitRequest": {
        "type": "object",
        "properties": {
          "transcriptId": { "type": "string", "format": "uuid" },
          "rawResponse": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000000
          },
          "save": { "type": "boolean" }
        },
        "required": ["transcriptId", "rawResponse"]
      },
      "GenerationResponse": {
        "type": "object",
        "properties": {
          "spec": { "type": "object" },
          "issues": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Issue" }
          },
          "attempts": { "type": "integer", "minimum": 1 }
        },
        "required": ["spec", "issues", "attempts"]
      },
      "ManualPromptResponse": {
        "type": "object",
        "properties": { "promptText": { "type": "string" } },
        "required": ["promptText"]
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "ComponentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "ValidationFailed": {
        "description": "Semantic validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/components": {
      "get": {
        "summary": "List all components",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Component registry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemsComponentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a component registry entry",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ComponentWriteRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created component",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Component" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/components/{id}": {
      "get": {
        "summary": "Get a component by id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ComponentId" }],
        "responses": {
          "200": {
            "description": "Component",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Component" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Update a component",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ComponentId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ComponentWriteRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated component",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Component" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Delete a component",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ComponentId" }],
        "responses": {
          "204": { "description": "Deleted" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/projects/{id}/generate": {
      "post": {
        "summary": "Generate a video spec using the configured AI provider",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spec generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenerationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{id}/generate/manual-prompt": {
      "post": {
        "summary": "Generate a manual prompt for external AI use",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ManualPromptRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt text generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualPromptResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/projects/{id}/generate/manual-submit": {
      "post": {
        "summary": "Submit a manually-generated JSON response",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ManualSubmitRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spec accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenerationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "422": { "$ref": "#/components/responses/ValidationFailed" }
        }
      }
    },
...
  }
}